% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPathway.R
\docType{package}
\name{gPathway}
\alias{gPathway}
\title{gPathway}
\usage{
gPathway(pathway, a, b)
}
\arguments{
\item{pathway}{A data frame. The columns must include "ID", "taxonomic.scope", and "Pathway".}

\item{a}{The output of \code{pre_process()}}

\item{b}{The output of \code{gClusters()}}
}
\description{
This function append the cluster information and the detailed feature ID to the pathway dataset.
}
\examples{
data("FuncExample")
head(a)
head(b)
ptw <- gPathway(pathway, a, b)
head(ptw)

}
